Web/backend/ml_models/gesture_recognition/inference.py
import torch
import cv2
import numpy as np
from PIL import Image

class GestureRecognitionModel:
    """Wrapper for gesture recognition ML models"""
    
    def __init__(self):
        self.model = None
        self.device = "cuda" if torch.cuda.is_available() else "cpu"
        self.class_names = [
            'bird', 'boar', 'dog', 'dragon', 'hare', 'horse',
            'monkey', 'ox', 'ram', 'rat', 'snake', 'tiger'
        ]
        # TODO: Initialize your YOLO/ViT models here
    
    def load_model(self, model_path: str):
        """Load pre-trained gesture recognition model"""
        # TODO: Load your YOLO/ViT model
        pass
    
    def predict(self, image: np.ndarray):
        """Predict hand sign from image"""
        
        # TODO: 1. Preprocess image for model input
        
        # TODO: 2. Run model inference
        
        # TODO: 3. Postprocess model outputs
        
        # Placeholder return
        return {
            "sign": "tiger",
            "confidence": 0.95,
            "all_predictions": {}
        }
    
    def _preprocess_image(self, image: np.ndarray):
        """Preprocess image for model input"""
        # TODO: Resize, normalize, convert format
        pass
    
    def _postprocess_predictions(self, model_output):
        """Convert model output to predictions"""
        # TODO: Apply softmax, get top predictions
        pass
