Web/backend/ml_models/style_transfer/inference.py
import torch
import numpy as np
from PIL import Image

class StyleTransferModel:
    """Wrapper for style transfer ML models"""
    
    def __init__(self):
        self.model = None
        self.device = "cuda" if torch.cuda.is_available() else "cpu"
        # TODO: Initialize your style transfer models here
    
    def load_model(self, model_path: str):
        """Load pre-trained style transfer model"""
        # TODO: Load your ArtFusion/SaMST/RLMiniStyler model
        pass
    
    def transfer_style(self, image: Image.Image, style: str, intensity: float = 0.8):
        """Apply style transfer to image"""
        
        # TODO: 1. Preprocess image for model input
        
        # TODO: 2. Run model inference
        
        # TODO: 3. Postprocess model output
        
        # TODO: 4. Return styled image
        
        return image  # Placeholder
    
    def _preprocess_image(self, image: Image.Image):
        """Preprocess image for model input"""
        # TODO: Resize, normalize, convert to tensor
        pass
    
    def _postprocess_output(self, model_output):
        """Convert model output back to image"""
        # TODO: Convert tensor to PIL Image
        pass
